chore: Fix potential issue with scripts updating the OpenAPI clients repos

+++ scripts/clients/build_override/git_push.sh
@@ -2,7 +2,7 @@
 # ref: https://help.github.com/articles/adding-an-existing-project-to-github-using-the-command-line/
 #
 # Export GIT_TOKEN for automatic login
-# Usage example: ./git_push.sh vcarluer Cosmo-Tech "minor update" "github.com"
+# Usage example: ./git_push.sh slashme Cosmo-Tech cosmotech-api-java-client "minor update" "github.com"
 set -e
 
 git_user_id=$1
@@ -12,7 +12,7 @@ release_note=$4
 git_host=$5
 
 if [ "$git_host" = "" ]; then
-    git_host="$GIT_HOST"
+    git_host="${GIT_HOST:-}"
     if [ "$git_host" = "" ]; then
         git_host="github.com"
     fi
@@ -20,49 +20,51 @@ if [ "$git_host" = "" ]; then
 fi
 
 if [ "$git_user_id" = "" ]; then
-    git_user_id="$GIT_USER_ID"
+    git_user_id="${GIT_USER_ID:-}"
     echo "[INFO] No command line input provided. Set \$git_user_id to $git_user_id"
 fi
 
 if [ "$git_organization_id" = "" ]; then
-    git_organization_id="$GIT_ORGANIZATION_ID"
+    git_organization_id="${GIT_ORGANIZATION_ID:-}"
     echo "[INFO] No command line input provided. Set \$git_organization_id to $git_organization_id"
 fi
 
 if [ "$git_repo_id" = "" ]; then
-    git_repo_id="$GIT_REPO_ID"
+    git_repo_id="${GIT_REPO_ID:-}"
     echo "[INFO] No command line input provided. Set \$git_repo_id to $git_repo_id"
 fi
 
 if [ "$release_note" = "" ]; then
-    release_note="$GIT_RELEASE_NOTE"
+    release_note="${GIT_RELEASE_NOTE:-}"
     if [ "$release_note" = "" ]; then
         release_note="Minor update"
     fi
     echo "[INFO] No command line input provided. Set \$release_note to $release_note"
 fi
 
+CURRENT_SCRIPT_DIR=$(realpath "$(dirname "$0")")
+
 # Create the release directory
-mkdir -p ../../../release
+WORKING_DIR=$(mktemp -d -t "${cosmotech-api-$1-client}"-XXXXX)
+echo "WORKING_DIR: ${WORKING_DIR}"
 
 # Sets the new remote URI
 if [ "$GIT_TOKEN" = "" ]; then
     echo "[INFO] \$GIT_TOKEN (environment variable) is not set. Using the git credential in your environment."
-    github_uri=https://${git_user_id}@${git_host}/${git_organization_id}/${git_repo_id}.git
+    github_uri="https://${git_user_id}@${git_host}/${git_organization_id}/${git_repo_id}.git"
 else
-    github_uri=https://${git_user_id}:${GIT_TOKEN}@${git_host}/${git_organization_id}/${git_repo_id}.git
+    github_uri="https://${git_user_id}:${GIT_TOKEN}@${git_host}/${git_organization_id}/${git_repo_id}.git"
 fi
 
 # Clone remote repository
-pushd ../../../release
-git clone ${github_uri}
+git clone "${github_uri}" "${WORKING_DIR}"
 # Delete all files to remove renamed or deleted files
-cd ${git_repo_id}
-rm -rf *
-popd
+rm -rf "${WORKING_DIR:?}/*"
+
 # Adds the files in the local repository
-cp -r ../* ../../../release/${git_repo_id}
-pushd ../../../release/${git_repo_id}
+cp -r "$(realpath "${CURRENT_SCRIPT_DIR}/..")"/* "${WORKING_DIR:?}"/
+
+pushd "${WORKING_DIR}" || exit 1
 
 git status
 
@@ -87,5 +89,4 @@ fi
 
 popd
 
-# Cleaning release repository
-rm -rf ../../../release
+rm -rf "${WORKING_DIR:?}" || exit 0
